patient_notes/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "patient_notes"
version = "0.0.1"
dependencies = []

[project.optional-dependencies]
build = [
    "build == 0.10.0"
]
test = [
    "delta-spark == 2.4.0",
    "opencensus-ext-azure==1.1.9",
    "opencensus-ext-logging==0.1.1",
    "pyodbc == 4.0.39",
    "pyspark == 3.4.1",
    "pytest == 7.2.1",
    "pytest-mock == 3.11.1",
    "pymssql == 2.2.8",
    "pyspark-test == 0.2.0", 
    "testcontainers == 3.7.1",
    "SQLAlchemy == 2.0.19",
    "presidio-analyzer == 2.2.33",
    "presidio-anonymizer == 2.2.33",
    "synapseml == 0.11.3",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl"

]
lint = [
    "mypy == 1.0.0",
    "pre-commit == 3.5.0"
]

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opencensus.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyspark_test.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "synapse.ml.cognitive.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "presidio_analyzer.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "presidio_anonymizer.*"
ignore_missing_imports = true
